+++ CHANGELOG.md
@@ -11,9 +11,11 @@ The source code to [eostoken.me](https://eostoken.me) adheres to [Semantic Versi
 
 - Generate [truffle-contract] sources from EOSSale and EOSToken artifacts.
 - Integrate [Web3].
+- Integrate [Redux] for managing state.
 
 [truffle-contract]: https://github.com/trufflesuite/truffle-contract
 [Web3]: https://github.com/ethereum/web3.js
+[Redux]: https://github.com/reactjs/redux
 
 ## [0.1.1][] - 2018-03-10